refatorando o codigo

+++ src/main/java/br/com/cirurgia/repository/CirurgiaRepository.java
@@ -3,4 +3,4 @@ package br.com.cirurgia.repository;
 import br.com.cirurgia.entity.Cirurgia;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CirurgiaRepository extends JpaRepository<Cirurgia, Integer> {}
+public interface CirurgiaRepository extends JpaRepository<Cirurgia, Long> {}
